requirements.txt
numpy==1.21.2
aiohttp==3.11.12
customtkinter==5.2.2
matplotlib==3.5.2
Pillow==11.2.1
ping3==4.0.8
psutil==5.9.1
pystray==0.19.5
python-dotenv==1.1.0
Requests==2.32.3
tenacity==8.2.2
